docker/day12_ConnectingAll/requirements.txt
fastapi[all]
pymongo
kafka-python

